feat: added create order service

+++ src/app/payment/page.tsx
@@ -1,23 +1,48 @@
 "use client";
-import React from "react";
+import React, { useState } from "react";
 import { useRouter } from "next/navigation";
 
 import useVelmarContext from "@/hooks/useVelmarContext";
 import { totalPrice } from "@/utils/functions-share";
 import { IoMdArrowRoundBack } from "react-icons/io";
+import { createOrder } from "@/services/orders.service";
+import { ICartItem } from "@/types/cart.types";
+import Loading from "@/components/Loading";
 
 const PaymentPage = () => {
   const router = useRouter();
 
   const { cart, deleteAllCart } = useVelmarContext();
 
+  const [loading, setLoading] = useState<boolean>(false);
+
   const handleBack = () => {
     router.back();
   };
 
-  const handlePay = () => {
-    router.push("/success");
-    deleteAllCart();
+  const prepareCart = (cart: ICartItem[]) => {
+    let cartFinal = [];
+
+    for (const item of cart) {
+      const { images, ...properties } = item;
+      cartFinal.push({ ...properties });
+    }
+
+    return cartFinal;
+  };
+
+  const handlePay = async () => {
+    try {
+      if (cart && cart.length > 0) {
+        await createOrder({ cart: prepareCart(cart) });
+        router.push("/success");
+        setLoading(true);
+        deleteAllCart();
+      }
+    } catch (error) {
+      console.error(error);
+      throw error;
+    }
   };
 
   return (
@@ -31,34 +56,42 @@ const PaymentPage = () => {
         <IoMdArrowRoundBack />
         Back
       </button>
-      <div className="bg-primary px-6 sm:px-12 py-12 text-secondary rounded-lg mb-12">
-        <h3 className="text-xl font-semibold">Order summary:</h3>
-        {cart?.map(({ id, name, price, quantity, size }) => (
-          <div className="border-b border-secondary m-4" key={id + size}>
-            <div className="flex flex-col">
-              <h4 className="font-semibold text-lg my-4">{name}</h4>
-              <span className="mb-4">Size: {size}</span>
-              <span className="mb-4">x {quantity}</span>
-              <span className="font-semibold">${price}</span>
-            </div>
+      <div className="bg-primary px-6 sm:px-12 py-12 text-secondary rounded-lg mb-12 min-h-80">
+        {loading ? (
+          <div className="grid place-content-center h-80">
+            <Loading />
           </div>
-        ))}
-        {cart && cart.length > 0 && (
+        ) : (
           <>
-            <br />
-            <div className="Payment-element">
-              <h3 className="text-xl font-semibold">Total to pay:</h3>
-              <h4 className="text-xl font-semibold">${totalPrice(cart)}</h4>
-            </div>
-            <br />
-            <button
-              type="button"
-              title="Pay"
-              className="p-3 mt-4 w-40 text-primary bg-secondary hover:bg-body hover:text-secondary hover:scale-105 duration-150"
-              onClick={handlePay}
-            >
-              Pay
-            </button>
+            <h3 className="text-xl font-semibold">Order summary:</h3>
+            {cart?.map(({ id, name, price, quantity, size }) => (
+              <div className="border-b border-secondary m-4" key={id + size}>
+                <div className="flex flex-col">
+                  <h4 className="font-semibold text-lg my-4">{name}</h4>
+                  <span className="mb-4">Size: {size}</span>
+                  <span className="mb-4">x {quantity}</span>
+                  <span className="font-semibold">${price}</span>
+                </div>
+              </div>
+            ))}
+            {cart && cart.length > 0 && (
+              <>
+                <br />
+                <div className="Payment-element">
+                  <h3 className="text-xl font-semibold">Total to pay:</h3>
+                  <h4 className="text-xl font-semibold">${totalPrice(cart)}</h4>
+                </div>
+                <br />
+                <button
+                  type="button"
+                  title="Pay"
+                  className="p-3 mt-4 w-40 text-primary bg-secondary hover:bg-body hover:text-secondary hover:scale-105 duration-150"
+                  onClick={handlePay}
+                >
+                  Pay
+                </button>
+              </>
+            )}
           </>
         )}
       </div>
